migrate to standard python-telegram-bot stuff
using Bot, Updater, Dispatcher and JobQueue from python-telegram-bot
