+++ main.py
@@ -1,8 +1,14 @@
 import logging
+
 import tweepy
 from envparse import Env
-from bot import TwitterForwarderBot, FetchAndSendTweetsJob
-from basebot import JobQueue
+from telegram.ext import CommandHandler
+from telegram.ext import Updater
+from telegram.ext.messagehandler import MessageHandler, Filters
+
+from bot import TwitterForwarderBot
+from commands import *
+from job import FetchAndSendTweetsJob
 
 env = Env(
     TWITTER_CONSUMER_KEY=str,
@@ -12,6 +18,7 @@ env = Env(
     TELEGRAM_BOT_TOKEN=str,
 )
 
+
 if __name__ == '__main__':
 
     logging.basicConfig(
@@ -19,9 +26,9 @@ if __name__ == '__main__':
         level=logging.WARNING)
 
     logging.getLogger(TwitterForwarderBot.__name__).setLevel(logging.DEBUG)
-    logging.getLogger(JobQueue.__name__).setLevel(logging.DEBUG)
     logging.getLogger(FetchAndSendTweetsJob.__name__).setLevel(logging.DEBUG)
 
+    # initialize Twitter API
     auth = tweepy.OAuthHandler(env('TWITTER_CONSUMER_KEY'), env('TWITTER_CONSUMER_SECRET'))
 
     try:
@@ -33,6 +40,27 @@ if __name__ == '__main__':
 
     twapi = tweepy.API(auth)
 
-    bot = TwitterForwarderBot(env('TELEGRAM_BOT_TOKEN'), twapi)
+    # initialize telegram API
+    token = env('TELEGRAM_BOT_TOKEN')
+    updater = Updater(bot=TwitterForwarderBot(token, twapi))
+    dispatcher = updater.dispatcher
 
-    bot.kb_interruptable_loop()
+    # set commands
+    dispatcher.add_handler(CommandHandler('start', cmd_start))
+    dispatcher.add_handler(CommandHandler('help', cmd_help))
+    dispatcher.add_handler(CommandHandler('ping', cmd_ping))
+    dispatcher.add_handler(CommandHandler('sub', cmd_sub, pass_args=True))
+    dispatcher.add_handler(CommandHandler('unsub', cmd_unsub, pass_args=True))
+    dispatcher.add_handler(CommandHandler('list', cmd_list))
+    dispatcher.add_handler(CommandHandler('export', cmd_export))
+    dispatcher.add_handler(CommandHandler('all', cmd_all))
+    dispatcher.add_handler(CommandHandler('wipe', cmd_wipe))
+    dispatcher.add_handler(CommandHandler('source', cmd_source))
+    dispatcher.add_handler(MessageHandler([Filters.text], handle_chat))
+
+    # put job
+    queue = updater.job_queue
+    queue.put(FetchAndSendTweetsJob(), next_t=0)
+
+    # poll
+    updater.start_polling()
