+++ util.py
@@ -1,0 +1,30 @@
+from functools import wraps
+import re
+
+
+def with_touched_chat(f):
+    @wraps(f)
+    def wrapper(bot, update=None, *args, **kwargs):
+        if update is None:
+            return f(bot, *args, **kwargs)
+
+        chat = bot.get_chat(update.message.chat)
+        chat.touch_contact()
+        kwargs.update(chat=chat)
+        return f(bot, update, *args, **kwargs)
+
+    return wrapper
+
+
+def escape_markdown(text):
+    """Helper function to escape telegram markup symbols"""
+    escape_chars = '\*_`\['
+    return re.sub(r'([%s])' % escape_chars, r'\\\1', text)
+
+
+def markdown_twitter_usernames(text):
+    """Restore markdown escaped usernames and make them link to twitter"""
+    return re.sub(r'@([^\s]*)',
+                  lambda s: '[@{username}](https://twitter.com/{username})'
+                  .format(username=s.group(1).replace(r'\_', '_')),
+                  text)
